Rename: 폴더 구조 변경
- app 폴더 내부는 라우팅용으로만 사용하도록 변경
