+++ src/app/contact/page.tsx
@@ -1,5 +1,5 @@
 import CommonFrame from "@/components/CommonFrame";
-import ContactContent from "./contact-content";
+import ContactContent from "@/container/contact/contact-content";
 
 export default function Contact() {
   return <CommonFrame sectionName="Contact" content={<ContactContent />} />;
