+++ src/app/feedback/page.tsx
@@ -1,5 +1,5 @@
 import CommonFrame from "@/components/CommonFrame";
-import FeedbackContent from "./feedback-content";
+import FeedbackContent from "@/container/feedback/feedback-content";
 
 export default function Feedback() {
   return <CommonFrame sectionName="Feedback" content={<FeedbackContent />} />;
