+++ src/app/projects/page.tsx
@@ -1,5 +1,5 @@
 import CommonFrame from "@/components/CommonFrame";
-import ProjectsContent from "./projects-content";
+import ProjectsContent from "@/container/projects/projects-content";
 
 export default function Projects() {
   return <CommonFrame sectionName="Projects" content={<ProjectsContent />} />;
