+++ src/container/projects/projects-content.tsx
@@ -1,7 +1,7 @@
 import styles from "@/styles/contents/projectscontent.module.css";
 
 import { projects } from "@/config/projects";
-import Project from "@/app/projects/project-form";
+import Project from "./project-form";
 
 export default function ProjectsContent() {
   return (
